moved atomics to common api
