+++ save-agent/save-agent-common-api/src/jsMain/kotlin/com/saveourtool/save/utils/PlatformUtils.kt
