+++ save-agent/save-agent-common/src/jvmMain/kotlin/com/saveourtool/save/utils/PlatformUtils.kt
@@ -7,15 +7,4 @@
 
 package com.saveourtool.save.utils
 
-actual typealias AtomicLong = java.util.concurrent.atomic.AtomicLong
-
-@Suppress("USE_DATA_CLASS")
-actual class GenericAtomicReference<T> actual constructor(valueToStore: T) {
-    private val holder: java.util.concurrent.atomic.AtomicReference<T> = java.util.concurrent.atomic.AtomicReference(valueToStore)
-    actual fun get(): T = holder.get()
-    actual fun set(newValue: T) {
-        holder.set(newValue)
-    }
-}
-
 actual fun getenv(envName: String): String? = System.getProperty(envName) ?: System.getenv(envName)
