+++ save-agent/save-demo-agent-api/build.gradle.kts
@@ -32,6 +32,7 @@ kotlin {
         val commonMain by getting {
             dependencies {
                 implementation(libs.save.common)
+                implementation(project(":save-agent:save-agent-common-api"))
                 api(libs.kotlinx.serialization.core)
                 api(libs.kotlinx.serialization.json)
                 api(libs.kotlinx.datetime)
@@ -42,7 +43,9 @@ kotlin {
                 implementation(libs.ktor.client.content.negotiation)
                 implementation(libs.ktor.serialization.kotlinx.json)
 
-                api(libs.cosv4k)
+                implementation(libs.ktoml.core)
+                implementation(libs.ktoml.source)
+                implementation(libs.ktoml.file)
             }
         }
         commonTest {
@@ -60,9 +63,6 @@ kotlin {
                 api(libs.slf4j.api)
                 api(libs.jetbrains.annotations)
                 implementation(libs.commons.compress)
-                implementation(libs.ktoml.core)
-                implementation(libs.ktoml.source)
-                implementation(libs.ktoml.file)
                 api(libs.kotlinx.coroutines.reactor)
             }
         }
@@ -76,20 +76,6 @@ kotlin {
                 api(libs.junit.jupiter.api)
                 api(libs.junit.jupiter.params)
                 runtimeOnly(libs.junit.jupiter.engine)
-            }
-        }
-
-        val linuxX64Main by getting
-
-        @Suppress("UNUSED_VARIABLE")
-        val nativeMain by creating {
-            dependsOn(commonMain)
-            linuxX64Main.dependsOn(this)
-
-            dependencies {
-                implementation(libs.ktoml.core)
-                implementation(libs.ktoml.source)
-                implementation(libs.ktoml.file)
             }
         }
     }
