+++ save-agent/save-demo-agent-api/src/commonMain/kotlin/com/saveourtool/save/utils/PlatformUtils.kt
@@ -11,45 +11,6 @@ import kotlin.time.DurationUnit
 import kotlinx.datetime.Clock
 
 /**
- * Atomic values
- */
-expect class AtomicLong(value: Long) {
-    /**
-     * @return value
-     */
-    fun get(): Long
-
-    /**
-     * @param newValue
-     */
-    fun set(newValue: Long)
-
-    /**
-     * @param delta increments the value_ by delta
-     * @return the new value
-     */
-    fun addAndGet(delta: Long): Long
-}
-
-/**
- *  Class that holds value and shares atomic reference to the value
- *
- *  @param valueToStore value to store
- */
-@Suppress("USE_DATA_CLASS")
-expect class GenericAtomicReference<T>(valueToStore: T) {
-    /**
-     * @return stored value
-     */
-    fun get(): T
-
-    /**
-     * @param newValue new value to store
-     */
-    fun set(newValue: T)
-}
-
-/**
  * A wrapper around a value of type [T] that caches it for [expirationTimeSeconds] and then recalculates
  * using [valueGetter]
  *
