+++ settings.gradle.kts
@@ -51,6 +51,7 @@ include("save-demo")
 include("save-demo-cpg")
 include("test-analysis-core")
 include(":save-agent:save-agent-common")
+include(":save-agent:save-agent-common-api")
 include(":save-agent:save-cloud-agent-api")
 include(":save-agent:save-cloud-agent")
 include(":save-agent:save-demo-agent-api")
